python/requirements.txt
joblib==1.4.2
numpy==2.0.0
pillow==10.3.0
scikit-learn==1.5.0
scipy==1.14.0
threadpoolctl==3.5.0
